Update can bo, hanh chinh
