+++ models/adBoardRequest.js
@@ -1,0 +1,13 @@
+import mongoose from "mongoose";
+const Schema = mongoose.Schema;
+import adBoard from "./adBoard.js";
+import User from "./user.js";
+
+const adBoardRequestSchema = new Schema({
+    adBoard: {type: Schema.Types.ObjectId, ref: "adBoard"},
+    sender: {type: Schema.Types.ObjectId, ref: "User"},
+    date: {type: Date},
+})
+
+const AdBoardReq = mongoose.model("AdBoardRequest", adBoardRequestSchema);
+export default AdBoardReq;
