+++ routes/so/hanhChinh.js
@@ -1,13 +1,16 @@
 import express from "express";
 const router = express.Router();
-import * as hanhChinh from '../../controllers/so/hanhChinh/hanhChinhControllers.js';
+import * as capPhepQC from '../../controllers/so/hanhChinh/capPhepQC.js';
+import * as chinhDiemQC from '../../controllers/so/hanhChinh/chinhDiemQC.js';
+import * as chinhBangQC from '../../controllers/so/hanhChinh/chinhBangQC.js';
+import * as thongKeBaoCao from '../../controllers/so/hanhChinh/thongKeBaoCao.js';
 
-router.get('/cap-phep-qc', hanhChinh.dsCapPhepQC);
-router.get('/cap-phep-qc/chi-tiet-yeu-cau', hanhChinh.chiTietYeuCauCapPhep);
-router.get('/chinh-diem-qc', hanhChinh.dsChinhDiemQC);
-router.get('/chinh-diem-qc/chi-tiet-yeu-cau', hanhChinh.chiTietChinhDiemQC);
-router.get('/chinh-bang-qc', hanhChinh.dsChinhBangQC);
-router.get('/chinh-bang-qc/chi-tiet-yeu-cau', hanhChinh.chiTietChinhBangQC);
-router.get('/thong-ke-qc', hanhChinh.tkBaoCaoXuLy);
+router.get('/cap-phep-qc', capPhepQC.dsCapPhepQC);
+router.get('/cap-phep-qc/chi-tiet-yeu-cau', capPhepQC.chiTietYeuCauCapPhep);
+router.get('/chinh-diem-qc', chinhDiemQC.dsChinhDiemQC);
+router.get('/chinh-diem-qc/chi-tiet-yeu-cau', chinhDiemQC.chiTietChinhDiemQC);
+router.get('/chinh-bang-qc', chinhBangQC.dsChinhBangQC);
+router.get('/chinh-bang-qc/chi-tiet-yeu-cau', chinhBangQC.chiTietChinhBangQC);
+router.get('/thong-ke-qc', thongKeBaoCao.tkBaoCaoXuLy);
 
 export default router;
