+++ views/so/canBo/chinhSuaTaiKhoan.ejs
@@ -17,20 +17,20 @@
 </style>
 
 <script>
-    // $(document).ready(function() {
-    //     $("#btnSubmit").click(function(){
-    //         var passInput=$("#password");
-    //         var passClass=$("#btnSubmit");
-    //         if(passInput.attr('type')==='password')
-    //         {
-    //             passInput.attr('type','text');
-    //             $("i", this).toggleClass("bi bi-eye bi bi-eye-slash");
-    //         } else {
-    //             passInput.attr('type','password');
-    //             $("i", this).toggleClass("bi bi-eye-slash bi bi-eye");
-    //         }
-    //     }); 
-    // });
+    $(document).ready(function() {
+        $("#btnSubmit").click(function(){
+            var passInput=$("#password");
+            var passClass=$("#btnSubmit");
+            if(passInput.attr('type')==='password')
+            {
+                passInput.attr('type','text');
+                $("i", this).toggleClass("bi bi-eye bi bi-eye-slash");
+            } else {
+                passInput.attr('type','password');
+                $("i", this).toggleClass("bi bi-eye-slash bi bi-eye");
+            }
+        }); 
+    });
 </script>
 
 <form action="/so/canbo/chinh-sua-tai-khoan-cb/<%= canBo._id %>?_method=PUT" method="POST" enctype="application/x-www-form-urlencoded">
@@ -57,11 +57,21 @@
                         <label class="fw-bold">Số điện thoại</label>
                         <input type="text" name="phone" placeholder="Nhập số điện thoại" value="<%= canBo.phone %>" class="form-control" required>
                     </div>
-                    <div class="col-md-10 mt-3 ps-3">
+                    <div class="col-md-5 mt-3 ps-3">
                         <label class="fw-bold">Email</label>
                         <input type="text" name="email" placeholder="Nhập địa chỉ email" value="<%= canBo.email %>" class="form-control" required>
                     </div>
-                    
+                    <div class="col-md-5 mt-3 ps-3">
+                        <label class="fw-bold">Vai trò</label>
+                        <select class="form-select" name="role" id="" required>
+                            <option selected>Cán bộ <%= canBo.role %></option>
+                            <% if (canBo.role == "quận") { %>
+                                <option value="phuong">Cán bộ phường</option>
+                            <% } else { %>
+                                <option value="quan">Cán bộ quận</option>
+                            <% } %>
+                        </select>
+                    </div>   
                     <div class="col-md-5 mt-3 ps-3">
                         <label class="fw-bold">Quận quản lý</label>
                         <select class="form-select" name="ward" id="" required>
@@ -93,11 +103,11 @@
                         <label class="fw-bold">Mật khẩu đăng nhập</label>
                         <div class="input-group">
                             <input type="text" name="password" id="password" placeholder="Nhập mật khẩu ban đầu cho cán bộ (thay đổi sau)" class="form-control" required>
-                            <!-- <span class="bg-white border-left-0">
-                                <button class="input-group-text bg-transparent" id="btnSubmit">
+                            <span class="bg-white border-left-0">
+                                <button class="input-group-text bg-transparent" id="btnSubmit" type="button">
                                     <i class="bi bi-eye-slash"></i>
                                 </button>
-                            </span> -->
+                            </span>
                         </div>
                     </div>
                     <div class="col-md-4 mt-5 mb-5 mx-1 mx-md-4 ps-3">
@@ -110,74 +120,6 @@
                 </div>
             </div>
         </div>
-        <div class="col-md-5 mt-3 ps-3">
-          <label class="fw-bold">Ngày sinh</label>
-          <input
-            type="date"
-            name=""
-            placeholder="Chọn ngày, tháng, năm sinh"
-            value="01/01/2001"
-            class="form-control"
-          />
-        </div>
-        <div class="col-md-5 mt-3 ps-3">
-          <label class="fw-bold">Số điện thoại</label>
-          <input
-            type="text"
-            name=""
-            placeholder="Nhập số điện thoại"
-            value="123456789"
-            class="form-control"
-          />
-        </div>
-        <div class="col-md-10 mt-3 ps-3">
-          <label class="fw-bold">Email</label>
-          <input
-            type="text"
-            name=""
-            placeholder="Nhập địa chỉ email"
-            value="nguyenvana@gmail.com"
-            class="form-control"
-          />
-        </div>
-
-        <div class="col-md-5 mt-3 ps-3">
-          <label class="fw-bold">Quận quản lý</label>
-          <select class="form-select" name="" id="">
-            <option selected>Chọn quận</option>
-            <option value="">Tân Bình</option>
-            <option value="">Quận 6</option>
-          </select>
-        </div>
-        <div class="col-md-5 mt-3 ps-3">
-          <label class="fw-bold">Phường quản lý</label>
-          <select class="form-select" name="" id="">
-            <option selected>Chọn phường</option>
-            <option value="">Phường 13</option>
-            <option value="">Phường 14</option>
-            <option value="">Phường 15</option>
-          </select>
-        </div>
-        <div class="col-md-10 mt-3 ps-3">
-          <label class="fw-bold">Mật khẩu đăng nhập</label>
-          <input
-            type="password"
-            name=""
-            placeholder=""
-            value="123"
-            class="form-control"
-          />
-        </div>
-        <div class="col-md-4 mt-5 mb-5 mx-1 mx-md-4 ps-3">
-          <button
-            class="form-control fs-6 fw-bold"
-            style="background-color: #468b4a"
-          >
-            <p class="text-white">Cập nhật tài khoản</p>
-          </button>
-        </div>
-      </div>
-    </div>
 </form>
 
 <div class="modal fade" tabindex="-1" role="dialog" id="deleteModal">
