+++ views/so/canBo/dkTaiKhoanCanBo.ejs
@@ -17,20 +17,20 @@
 </style>
 
 <script>
-    // $(document).ready(function() {
-    //     $("#btnSubmit").click(function(){
-    //         var passInput=$("#password");
-    //         var passClass=$("#btnSubmit");
-    //         if(passInput.attr('type')==='password')
-    //         {
-    //             passInput.attr('type','text');
-    //             $("i", this).toggleClass("bi bi-eye bi bi-eye-slash");
-    //         } else {
-    //             passInput.attr('type','password');
-    //             $("i", this).toggleClass("bi bi-eye-slash bi bi-eye");
-    //         }
-    //     }); 
-    // });
+    $(document).ready(function() {
+        $("#btnSubmit").click(function(){
+            var passInput=$("#password");
+            var passClass=$("#btnSubmit");
+            if(passInput.attr('type')==='password')
+            {
+                passInput.attr('type','text');
+                $("i", this).toggleClass("bi bi-eye bi bi-eye-slash");
+            } else {
+                passInput.attr('type','password');
+                $("i", this).toggleClass("bi bi-eye-slash bi bi-eye");
+            }
+        }); 
+    });
 </script>
 
 <form action="/so/canbo/dang-ky-tai-khoan-cb" method="POST">
@@ -50,17 +50,24 @@
                 </div>
                 <div class="col-md-5 mt-3 ps-3">
                     <label class="fw-bold">Ngày sinh</label>
-                    <input type="date" name="birthday" placeholder="Chọn ngày, tháng, năm sinh" class="form-control" required>
+                    <input type="date" name="birthday" placeholder="Chọn ngày, tháng, năm sinh" value="2018-01-01" class="form-control" required>
                 </div>
                 <div class="col-md-5 mt-3 ps-3">
                     <label class="fw-bold">Số điện thoại</label>
                     <input type="text" name="phone" placeholder="Nhập số điện thoại" class="form-control" required>
                 </div>
-                <div class="col-md-10 mt-3 ps-3">
+                <div class="col-md-5 mt-3 ps-3">
                     <label class="fw-bold">Email</label>
-                    <input type="text" name="email" placeholder="Nhập địa chỉ email" class="form-control" required>
+                    <input type="email" name="email" placeholder="Nhập địa chỉ email" class="form-control" required>
                 </div>
-                
+                <div class="col-md-5 mt-3 ps-3">
+                    <label class="fw-bold">Vai trò</label>
+                    <select class="form-select" name="role" id="" required>
+                        <option selected>Chọn vai trò</option>
+                        <option value="phường">Cán bộ phường</option>
+                        <option value="quận">Cán bộ quận</option>
+                    </select>
+                </div>   
                 <div class="col-md-5 mt-3 ps-3">
                     <label class="fw-bold">Quận quản lý</label>
                     <select class="form-select" name="ward" id="ward" required>
@@ -82,11 +89,11 @@
                     <label class="fw-bold">Mật khẩu đăng nhập</label>
                     <div class="input-group">
                         <input type="text" name="password" id="password" placeholder="Nhập mật khẩu ban đầu cho cán bộ (thay đổi sau)" class="form-control" required>
-                        <!-- <span class="bg-white border-left-0">
-                            <button class="input-group-text bg-transparent" id="btnSubmit">
+                        <span class="bg-white border-left-0">
+                            <button class="input-group-text bg-transparent" id="btnSubmit" type="button">
                                 <i class="bi bi-eye-slash"></i>
                             </button>
-                        </span> -->
+                        </span>
                     </div>
                 </div>
                 <div class="col-md-4 mt-5 mb-5 mx-1 mx-md-4 ps-3">
