+++ views/so/canBo/dsTaiKhoanCanBo.ejs
@@ -85,7 +85,7 @@
       </thead>
       <tbody>
         <% for (let i=0; i<canBo.length; i++) { %>
-          <% if (canBo[i].role == "quan" ) { %>
+          <% if (canBo[i].role == "Quận" || canBo[i].role == "Phường") { %>
             <tr class="align-items-center">
               <th scope="row">
                 <%= i + 1 %>
