+++ views/so/hanhChinh/chiTiet/ndChinhDiemQC.ejs
@@ -85,28 +85,8 @@
             </div>
         </li>
     </ul>
-    <ul class="list-group mt-5">
-        <li class="list-group-item" style="background-color: #7981d0">
-            <div class="row">
-                <div class="col" style="max-width: 200px"><b class="table header">Xử lý yêu cầu</b></div>
-            </div>
-        </li>
-        <li class="list-group-item">
-            <div class="row">
-                <div class="col header"><b>Thực hiện</b></div>
-                    <div class="col">
-                        <input type="radio" class="btn-check" style="width: 100%;" name="options-outlined" id="success-outlined" autocomplete="off" checked>
-                        <label class="btn btn-outline-success" style="width: 100%;" for="success-outlined">Phê duyệt</label>
-                    </div>
-                    <div class="col">
-                        <input type="radio" class="btn-check" style="width: 100%;" name="options-outlined" id="danger-outlined" autocomplete="off">
-                        <label class="btn btn-outline-danger" style="width: 100%;" for="danger-outlined">Không phê duyệt</label>
-                    </div>
-                    <div class="col">
-                        <button type="submit" class="btn"style="background-color: #D9DBE4; width: 100%;"><i class="bi bi-send-fill"></i>Gửi</button>
-                    </div>
-                </div>
-            </div>
-        </li>
-    </ul>
+    <div class="border-bottom border-top mt-5 mb-5" style="display: flex; justify-content: center; padding-top: 1rem; padding-bottom: 1rem;">
+        <a class="btn btn-danger mx-1" style="width: 25%;" for="success-outlined">Không phê duyệt</a>
+        <a class="btn btn-success mx-1" style="width: 25%;;" for="success-outlined">Phê duyệt</a>
+    </div>
 </div>
